requirements.txt
pandas==3.0.0
protobuf==5.28.3
tensorflow-cpu==2.20.0
